requirements.txt
flask==2.3.3
flask-cors==4.0.0
mysql-connector-python==8.1.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
python-dotenv==1.0.0
requests==2.31.0 